test: 수정된 내용을 반영하여 테스트 코드 재작성
